applesingle: Gave up on tagged unions for now.
Switched to enum values for each entry type tag.
Also, added some helper methods for extracting the filename, finder info,
resource, and data forks from the iospace.

Not sure whether it is worth doing the resource/data forks as byte arrays since
it is really slow. Need to look into switching to a better representation of the
offset, length, and data type values.