applesingle-extract: Added generic extraction.
Unfortunately for now you need to enter the ID numerically like:

applesingle-extract file.as 2

to get the resource fork if it exists or use 1 to get the data fork, etc.

Hopefully it will be easy to create an argument parser in Poke that can handle
int values or string symbols that map back to the reserved entry type values.