applesingle-combine: Re-vamped code.
The big new feature is autodetection of custom icon based on presence of `icns`
resource in resource fork.

Otherwise, made use of the poke stdlib argp module to parse arguments. I used
the handler for each option to populate the CLI invocation struct.

Also, added an encapsulation of input/output file arguments to make it easier to
copy an input iospace to the output file.