applersrc: Switched RsrcFork to relative offsets.
This allows the fork to be built from a section inside the currently mapped
iospace rather than a thing that begins at the start of the iospace so now it
should be possible to shift over to the resource fork in an applesingle file.

Also, removed the NameList struct because it was not very useful and typed a
few offset values as offset types.